test-requirements.txt
-r requirements.txt
pytest
pytest-cov
mock
PyHamcrest==1.8.0